requirements.txt
# Root project dependencies for development and testing
boto3>=1.34.0
botocore>=1.34.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
types-boto3>=1.0.0
jsonschema>=4.20.0
pyyaml>=6.0.0
mcp>=1.0.0
strand-sdk>=1.0.0